rename: ConsoleReader, RandomNumberGenerator Util 패키지로 이동
